+++ src/main/kotlin/be/vreijsenj/aoc/utils/Cube.kt
@@ -1,6 +1,6 @@
 package be.vreijsenj.aoc.utils
 
-open class Cube(open val x: Long, open val y: Long, open val z: Long) {
+open class Cube(open val x: Double, open val y: Double, open val z: Double) {
 
     fun above() = Cube(x, y, z + 1)
     fun below() = Cube(x, y, z - 1)
